tracker: fix tests (perhaps)
