+++ tracker/server_test.go
@@ -2,14 +2,20 @@ package tracker
 
 import (
 	"context"
+	"fmt"
+	"math/rand/v2"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"strings"
+	"sync"
+	"sync/atomic"
 	"testing"
+	"time"
 
 	radio "github.com/R-a-dio/valkyrie"
 	"github.com/R-a-dio/valkyrie/config"
+	"github.com/R-a-dio/valkyrie/mocks"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -19,10 +25,9 @@ func TestListenerAddAndRemove(t *testing.T) {
 	defer cancel()
 	cfg := config.TestConfig()
 
-	recorder := NewRecorder(ctx, cfg)
-	dummy := NewServer(ctx, "", recorder)
+	dummy := NewServer(ctx, cfg)
 
-	srv := httptest.NewServer(dummy.Handler)
+	srv := httptest.NewServer(dummy.h)
 	defer srv.Close()
 	client := srv.Client()
 
@@ -45,9 +50,9 @@ func TestListenerAddAndRemove(t *testing.T) {
 
 		// we also should have a listener in the recorder
 		require.Eventually(t, func() bool {
-			return assert.Equal(t, int64(1), recorder.ListenerAmount())
+			return assert.Equal(t, int64(1), dummy.recorder.ListenerAmount())
 		}, eventuallyDelay, eventuallyTick)
-		testListenerLength(t, recorder, 1)
+		testListenerLength(t, dummy.recorder, 1)
 
 		// =========================
 		// Do a normal leave request
@@ -62,10 +67,10 @@ func TestListenerAddAndRemove(t *testing.T) {
 		require.Equal(t, http.StatusOK, resp.StatusCode)
 		// and the listener should now be gone
 		require.Eventually(t, func() bool {
-			return assert.Equal(t, int64(0), recorder.ListenerAmount())
+			return assert.Equal(t, int64(0), dummy.recorder.ListenerAmount())
 		}, eventuallyDelay, eventuallyTick)
 
-		testListenerLength(t, recorder, 0)
+		testListenerLength(t, dummy.recorder, 0)
 	})
 
 	for _, uri := range []string{"/listener_joined", "/listener_left"} {
@@ -123,5 +128,71 @@ func BenchmarkListenerAdd(b *testing.B) {
 
 	for i := 0; i < b.N; i++ {
 		handler.ServeHTTP(w, req)
+	}
+}
+
+func TestPeriodicallyUpdateListeners(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	cfg := config.TestConfig()
+
+	done := make(chan struct{})
+
+	dummy := new(Server)
+	dummy.cfg = cfg
+	dummy.recorder = NewRecorder(ctx, cfg)
+
+	var last atomic.Int64
+	var count int
+	var closeOnce sync.Once
+
+	manager := &mocks.ManagerServiceMock{
+		UpdateListenersFunc: func(contextMoqParam context.Context, new int64) error {
+			// we're done after 10 updates
+			if count++; count > 10 {
+				closeOnce.Do(func() {
+					close(done)
+				})
+				return nil
+			}
+			// every 5 updates return an error
+			if count%5 == 0 {
+				return fmt.Errorf("that's an error")
+			}
+
+			// otherwise our new value should equal what we set it to previously
+			if !assert.Equal(t, last.Load(), new) {
+				closeOnce.Do(func() {
+					close(done)
+				})
+				return nil
+			}
+
+			adjustment := rand.Int64()
+			dummy.recorder.listenerAmount.Store(adjustment)
+			last.Store(adjustment)
+
+			return nil
+		},
+	}
+	dummy.cfg.Manager = manager
+
+	finished := make(chan struct{})
+	go func() {
+		defer close(finished)
+		dummy.periodicallyUpdateListeners(ctx, time.Millisecond*10)
+	}()
+
+	// wait for the 10 updates
+	<-done
+
+	// cancel the context we gave the function, it should clean
+	// itself up
+	cancel()
+
+	select {
+	case <-finished:
+	case <-time.After(eventuallyDelay):
+		t.Error("failed to cleanup")
 	}
 }
